requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.44.0
streamlit>=1.31.0
plotly>=5.18.0
pydantic>=2.5.0
python-dotenv>=1.0.0
groq>=0.4.0
pytest>=7.4.0

